Refactor package structure and improve Card number validation class
Moved bank-related services from 'com.epam.engx.jmp.bankapi' to 'com.epam.engx.jmp.api' to standardize the package structure. Renamed 'LuhnAlgorithmCheck' to 'CardNumberValidator' for more clarity and extended its functionality to reject Card numbers not conforming to basic card number formality along with Luhn algorithm validation. Updated tests and dependencies to reflect these changes. Updated project configuration for better integration with SonarCloud and adjusted reporting in pom.xml for nuanced test insights.
