+++ .github/workflows/build.yml
@@ -44,4 +44,7 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}  # Needed to get PR information, if any
           SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
-        run: mvn -B verify org.sonarsource.scanner.maven:sonar-maven-plugin:sonar -Dsonar.projectKey=rabestro_course-java9-modules
+        run: mvn verify org.sonarsource.scanner.maven:sonar-maven-plugin:sonar
+          --batch-mode
+          --define sonar.host.url=https://sonarcloud.io
+          --define sonar.projectKey=org.epam.engx:jmp
