+++ jmp-cloud-bank-impl/src/main/java/module-info.java
@@ -1,5 +1,8 @@
+import com.epam.engx.jmp.api.Bank;
+import com.epam.engx.jmp.cloud.bank.impl.BankCloudImpl;
+
 module jmp.cloud.bank.impl {
 	requires transitive jmp.bank.api;
 	requires jmp.dto;
-	exports com.epam.engx.jmp.cloud.bank.impl;
+	provides Bank with BankCloudImpl;
 }
