+++ jmp-cloud-bank-impl/src/test/groovy/com/epam/engx/jmp/cloud/bank/impl/CardNumberValidatorSpec.groovy
@@ -2,11 +2,11 @@ package com.epam.engx.jmp.cloud.bank.impl
 
 import spock.lang.Specification
 
-class LuhnAlgorithmCheckSpec extends Specification {
+class CardNumberValidatorSpec extends Specification {
 
 	def "Test with a known valid card number: #cardNumber"() {
 		given:
-		def luhnCheck = new LuhnAlgorithmCheck()
+		def luhnCheck = new CardNumberValidator()
 
 		when:
 		def isValid = luhnCheck.test(cardNumber)
@@ -26,7 +26,7 @@ class LuhnAlgorithmCheckSpec extends Specification {
 
 	def "Test with a known invalid card number: #cardNumber"() {
 		given:
-		def luhnCheck = new LuhnAlgorithmCheck()
+		def luhnCheck = new CardNumberValidator()
 
 		when:
 		def isValid = luhnCheck.test(cardNumber)
@@ -46,7 +46,7 @@ class LuhnAlgorithmCheckSpec extends Specification {
 
 	def "Test with a null card number"() {
 		given:
-		def luhnCheck = new LuhnAlgorithmCheck()
+		def luhnCheck = new CardNumberValidator()
 
 		when:
 		def isValid = luhnCheck.test(null)
@@ -57,7 +57,7 @@ class LuhnAlgorithmCheckSpec extends Specification {
 
 	def "Test with an empty card number"() {
 		given:
-		def luhnCheck = new LuhnAlgorithmCheck()
+		def luhnCheck = new CardNumberValidator()
 
 		when:
 		def isValid = luhnCheck.test("")
@@ -68,7 +68,7 @@ class LuhnAlgorithmCheckSpec extends Specification {
 
 	def "Test with a card number containing non-numeric characters"() {
 		given:
-		def luhnCheck = new LuhnAlgorithmCheck()
+		def luhnCheck = new CardNumberValidator()
 
 		when:
 		def isValid = luhnCheck.test("4532a15112830366")
